Fixes #117 Split battle files and clarify engagement pipeline call flow
