+++ src/game/engagement/battle/tick.rs
@@ -14,7 +14,28 @@ use super::{
     BattleMessage, BattlePhase, BattleTick, QueuedAbility, entity_innate_battle_abilities,
 };
 
-pub async fn handle_tick(game_state: &Arc<GameState>, result: BattleTick, max_engage_ticks: u64) {
+/// The outcome of resolving a single battle tick. Returned to the caller so that engagement
+/// lifecycle operations (removing dead participants, concluding, removing the engagement) happen
+/// in one place rather than being scattered across the resolution pipeline.
+pub struct BattleTickOutcome {
+    pub engagement_id: i64,
+    /// All participant ids at the start of this tick (before any deaths are processed).
+    pub all_participant_ids: Vec<i64>,
+    /// Entity ids whose HP dropped to or below their minimum this tick.
+    pub dead_entity_ids: Vec<i64>,
+    /// Player ids subscribed to this battle (for follow-up messaging).
+    pub player_ids: Vec<i64>,
+}
+
+/// Resolves effects from a completed battle tick: applies innate and queued ability effects,
+/// detects entity deaths, and broadcasts the battle state update to all player participants.
+/// Returns a `BattleTickOutcome` so the caller can apply engagement lifecycle changes
+/// (update participants, conclude, remove) without this function reaching back into `Engagements`.
+pub async fn handle_tick(
+    game_state: &Arc<GameState>,
+    result: BattleTick,
+    max_engage_ticks: u64,
+) -> BattleTickOutcome {
     let engagement_id = result.engagement_id;
     let all_ids = result.all_participant_ids.clone();
 
@@ -54,11 +75,6 @@ pub async fn handle_tick(game_state: &Arc<GameState>, result: BattleTick, max_en
 
     let player_ids = find_participant_player_ids(game_state, &all_ids).await;
 
-    let surviving = game_state
-        .engagements
-        .update_battle_participants(engagement_id, &dead_ids)
-        .await;
-
     let countdown_ticks = max_engage_ticks.saturating_sub(result.ticks_in_phase);
     let params = BattleUpdateParams {
         engagement_id,
@@ -75,14 +91,21 @@ pub async fn handle_tick(game_state: &Arc<GameState>, result: BattleTick, max_en
         messaging::battle_update(&game_state.message_tx, pid, update.clone());
     }
 
-    if surviving <= 1 {
-        loot::resolve_loot(&all_ids);
-        game_state.engagements.conclude_battle(engagement_id).await;
-        clear_active_effects(game_state, &all_ids).await;
-        for &pid in &player_ids {
-            messaging::battle_ended(&game_state.message_tx, pid, engagement_id);
-        }
-        game_state.engagements.remove(engagement_id).await;
+    BattleTickOutcome {
+        engagement_id,
+        all_participant_ids: all_ids,
+        dead_entity_ids: dead_ids,
+        player_ids,
+    }
+}
+
+/// Called when a battle ends (surviving factions <= 1). Clears transient effects from all
+/// participants and sends the battle-ended event to each player.
+pub async fn handle_battle_ended(game_state: &Arc<GameState>, outcome: &BattleTickOutcome) {
+    loot::resolve_loot(&outcome.all_participant_ids);
+    clear_active_effects(game_state, &outcome.all_participant_ids).await;
+    for &pid in &outcome.player_ids {
+        messaging::battle_ended(&game_state.message_tx, pid, outcome.engagement_id);
     }
 }
